feat: 文件分片合并

+++ server/web/src/main/java/com/xiaohe/pan/server/web/model/domain/FileChunk.java
@@ -74,7 +74,7 @@ public class FileChunk extends BaseDomain {
     /**
      * 分片文件的大小
      */
-    private Integer chunkSize;
+    private Long chunkSize;
 
     /**
      * 分片所属文件的类型
